refactor: rename UserSettings component and update imports
